updated sys::recurseRmDir() - added check if folder exists

+++ system/classes/sys.php
@@ -1176,11 +1176,18 @@ namespace YAWK {
          * @return bool
          */
         static function recurseRmdir($dir) {
-            $files = array_diff(scandir($dir), array('.','..'));
-            foreach ($files as $file) {
-                (is_dir("$dir/$file")) ? self::recurseRmdir("$dir/$file") : unlink("$dir/$file");
+            if (is_dir(dirname($dir)))
+            {
+                $files = array_diff(scandir($dir), array('.','..'));
+                foreach ($files as $file) {
+                    (is_dir("$dir/$file")) ? self::recurseRmdir("$dir/$file") : unlink("$dir/$file");
+                }
+                return rmdir($dir);
             }
-            return rmdir($dir);
+            else
+                {
+                    return false;
+                }
         }
 
         /**
